[MNG-7789] Dependency validation rules used wrong data
Refactored and now split, we have "plugin dependency" (using POM)
and "plugin descriptor" (using PluginDescriptor/dependencies) to
perform validation.

Also, two "inlined" validation from plugin dependencies resolver
factored out to components, along with 3 existing moved to newly
added (moved) validators.

---

https://issues.apache.org/jira/browse/MNG-7789
