+++ maven-core/src/main/java/org/apache/maven/plugin/internal/MavenMixedDependenciesValidator.java
@@ -25,10 +25,11 @@ import javax.inject.Singleton;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.PluginValidationManager;
-import org.apache.maven.plugin.descriptor.MojoDescriptor;
-import org.codehaus.plexus.component.repository.ComponentDependency;
+import org.eclipse.aether.RepositorySystemSession;
+import org.eclipse.aether.artifact.Artifact;
+import org.eclipse.aether.graph.Dependency;
+import org.eclipse.aether.resolution.ArtifactDescriptorResult;
 
 /**
  * Detects mixed Maven versions in plugins.
@@ -45,16 +46,21 @@ class MavenMixedDependenciesValidator extends AbstractMavenPluginDependenciesVal
     }
 
     @Override
-    protected void doValidate(MavenSession mavenSession, MojoDescriptor mojoDescriptor) {
-        Set<String> mavenVersions = mojoDescriptor.getPluginDescriptor().getDependencies().stream()
+    protected void doValidate(
+            RepositorySystemSession session,
+            Artifact pluginArtifact,
+            ArtifactDescriptorResult artifactDescriptorResult) {
+        Set<String> mavenVersions = artifactDescriptorResult.getDependencies().stream()
+                .map(Dependency::getArtifact)
                 .filter(d -> "org.apache.maven".equals(d.getGroupId()))
-                .filter(d -> !EXPECTED_PROVIDED_SCOPE_EXCLUSIONS_GA.contains(d.getGroupId() + ":" + d.getArtifactId()))
-                .map(ComponentDependency::getVersion)
+                .filter(d -> !DefaultPluginValidationManager.EXPECTED_PROVIDED_SCOPE_EXCLUSIONS_GA.contains(
+                        d.getGroupId() + ":" + d.getArtifactId()))
+                .map(Artifact::getVersion)
                 .collect(Collectors.toSet());
 
         if (mavenVersions.size() > 1) {
             pluginValidationManager.reportPluginValidationIssue(
-                    mavenSession, mojoDescriptor, "Plugin mixes multiple Maven versions: " + mavenVersions);
+                    session, pluginArtifact, "Plugin mixes multiple Maven versions: " + mavenVersions);
         }
     }
 }
